Account for missing images
If we're requesting an index beyond what we have, we missed an image or more.  Add the "no image" into the array as necessary

Note: when setting the maximum of the slider, use the max of the number of images in either array.  We have no way of removing extra images