Sync code improvements (again)
Always keep the array lengths in sync.  When adding a new image, check if the last image was a dummy (ie timestamp == 0) and replace it if this new image is +/-2s different from what the greatest previous timestamp was

Assumes that the cameras are in sync time-wise and that the cameras will alternate sending the new images.  If they don't alternate sending images (ie camera 1 sends 3 images, and then camera 2 sends 3 images and there are only three distinct triggers, then we will end up with 5 entries)