Better handle different widths
Previously, we had been hardcoding the width of the images as 500px as this is approximately what looks best when the screen is 1366px wide.
However, it doesn't work well on lower/higher resolutions.  Make the images adjust to the total width of the monitor in pixels - not to the screen size as the app is next to useless when not fullscreen